requirements.txt
tensorflow
fastapi
uvicorn
pillow
requests
python-multipart 